Major refactor of the tests.
- It is now easier to follow what the tests do.
- Service tests added for SNMP (as the basis for the rest).
- Fixed issues with commands that expect Int/Int64 attributes.
