+++ cli/provisioning/foreign_sources_test.go
@@ -11,7 +11,8 @@ import (
 
 func TestGetForeignSource(t *testing.T) {
 	var err error
-	app, server := test.InitializeMocks(t, ForeignSourcesCliCommand)
+	app := test.CreateCli(ForeignSourcesCliCommand)
+	server := createTestServer(t)
 	defer server.Close()
 
 	err = app.Run([]string{app.Name, "fs", "get"})
@@ -23,7 +24,8 @@ func TestGetForeignSource(t *testing.T) {
 
 func TestSetScanInterval(t *testing.T) {
 	var err error
-	app, server := test.InitializeMocks(t, ForeignSourcesCliCommand)
+	app := test.CreateCli(ForeignSourcesCliCommand)
+	server := createTestServer(t)
 	defer server.Close()
 
 	err = app.Run([]string{app.Name, "fs", "int"})
@@ -41,7 +43,8 @@ func TestSetScanInterval(t *testing.T) {
 
 func TestDeleteForeignSource(t *testing.T) {
 	var err error
-	app, server := test.InitializeMocks(t, ForeignSourcesCliCommand)
+	app := test.CreateCli(ForeignSourcesCliCommand)
+	server := createTestServer(t)
 	defer server.Close()
 
 	err = app.Run([]string{app.Name, "fs", "del"})
@@ -53,7 +56,8 @@ func TestDeleteForeignSource(t *testing.T) {
 
 func TestApplyForeignSource(t *testing.T) {
 	var err error
-	app, server := test.InitializeMocks(t, ForeignSourcesCliCommand)
+	app := test.CreateCli(ForeignSourcesCliCommand)
+	server := createTestServer(t)
 	defer server.Close()
 
 	err = app.Run([]string{app.Name, "fs", "apply"})
