+++ go.mod
@@ -3,7 +3,7 @@ module github.com/OpenNMS/onmsctl
 go 1.12
 
 require (
-	github.com/google/go-cmp v0.3.0 // indirect
+	github.com/google/go-cmp v0.3.0
 	github.com/pkg/errors v0.8.1 // indirect
 	github.com/urfave/cli v1.21.0
 	gopkg.in/yaml.v2 v2.2.2
